+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix for tooltip being hidden when CTRL key is pressed to move drag handle precisely
 - Fix for drag handle not adapting to magnetic selection on release
 - Fix for textfield buttons borders in 'reverse buttons order' mode
+- Added fallback url to fetch custom search buttons favicons
 
 #3.0.6
 - Implemented live translation of selected text (up to 3 words)